HS-192: Add metrics api & integreate it with minion RPC/heartbeat
