+++ platform/assemblies/features/pom.xml
@@ -84,12 +84,13 @@
                                 <feature>horizon-events</feature>
                                 <feature>horizon-grpc</feature>
                                 <feature>horizon-icmp</feature>
+                                <feature>horizon-inventory</feature>
                                 <feature>horizon-ipc-api</feature>
                                 <feature>horizon-provision</feature>
+                                <feature>horizon-minion-heartbeat</feature>
+                                <feature>horizon-minion-rest</feature>
                                 <feature>horizon-snmp</feature>
                                 <feature>horizon-tracing</feature>
-                                <feature>horizon-minion-rest</feature>
-                                <feature>horizon-inventory</feature>
                                 <feature>jackson</feature>
                                 <feature>java-native-access</feature>
                                 <feature>kie-drools</feature>
@@ -101,6 +102,7 @@
                                 <feature>camel-jaxb</feature>
                                 <feature>camel-kafka</feature>
                                 <feature>camel-activemq</feature>
+                                <feature>horizon-metrics-prometheus</feature>
                             </features>
                         </configuration>
                     </execution>
@@ -392,6 +394,11 @@
         <dependency>
             <groupId>org.opennms.horizon.minion</groupId>
             <artifactId>org.opennms.horizon.minion.rest</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.opennms.horizon.minion</groupId>
+            <artifactId>heartbeat</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
