--- platform/ipc/heartbeat/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,21 +1,0 @@
-<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0"
-           xsi:schemaLocation="
-		http://www.osgi.org/xmlns/blueprint/v1.0.0
-		https://osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
-		http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0
-		http://aries.apache.org/schemas/blueprint-cm/blueprint-cm-1.3.0.xsd
-">
-
-    <reference id="messageConsumerManager" interface="org.opennms.horizon.ipc.sink.api.MessageConsumerManager"/>
-    <reference id="minionDao" interface="org.opennms.horizon.db.dao.api.MinionDao"/>
-    <reference id="sessionUtils" interface="org.opennms.horizon.db.dao.api.SessionUtils" />
-
-    <bean id="heartbeatConsumer" class="org.opennms.ipc.heartbeat.HeartbeatConsumer" init-method="init">
-        <argument ref="messageConsumerManager"/>
-        <argument ref="minionDao"/>
-        <argument ref="sessionUtils"/>
-    </bean>
-
-</blueprint>
