+++ platform/minion/heartbeat/pom.xml
@@ -1,18 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
+    <parent>
+        <artifactId>org.opennms.horizon.minion</artifactId>
+        <groupId>org.opennms.horizon</groupId>
+        <version>0.1.0-SNAPSHOT</version>
+    </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <parent>
-        <groupId>org.opennms.horizon.ipc</groupId>
-        <artifactId>ipc</artifactId>
-        <version>0.1.0-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-
-    <artifactId>echo</artifactId>
-    <name>OpenNMS Horizon :: IPC :: Echo</name>
+    <groupId>org.opennms.horizon.minion</groupId>
+    <artifactId>heartbeat</artifactId>
+    <name>OpenNMS Horizon :: Minion :: Heartbeat</name>
     <packaging>bundle</packaging>
 
     <dependencyManagement>
@@ -34,13 +33,18 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
+            <groupId>org.opennms.horizon.db</groupId>
+            <artifactId>dao-api</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>org.opennms.horizon.ipc</groupId>
             <artifactId>api</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.opennms.horizon.metrics</groupId>
-            <artifactId>api</artifactId>
+            <groupId>org.opennms.horizon.core</groupId>
+            <artifactId>identity</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
@@ -48,11 +52,22 @@
             <artifactId>slf4j-api</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.osgi</groupId>
-            <artifactId>osgi.core</artifactId>
+            <groupId>org.opennms.horizon.metrics</groupId>
+            <artifactId>api</artifactId>
+            <version>${project.version}</version>
         </dependency>
-    </dependencies>
+        <!--test dependencies -->
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-core</artifactId>
+            <version>${mockito.version}</version>
+            <scope>test</scope>
+        </dependency>
 
-    <build>
-    </build>
+    </dependencies>
 </project>
