+++ platform/pom.xml
@@ -29,6 +29,7 @@
         <module>docker-it</module>
         <module>inventory</module>
         <module>minion</module>
+        <module>metrics</module>
     </modules>
 
     <properties>
@@ -95,6 +96,7 @@
         <osgi.version>7.0.0</osgi.version>
         <pax.jdbc.version>1.5.2</pax.jdbc.version>
         <postgresql.version>42.3.2</postgresql.version>
+        <prometheus.version>0.16.0</prometheus.version>
         <protobuf.version>3.16.1</protobuf.version>
         <rate.limitted.logger.version>2.0.2</rate.limitted.logger.version>
         <rest-assured.version>4.3.3</rest-assured.version>
